Move saving of a workbook part from XLWorkbook_Save to a separate writer.
